+++ controllers/dnschaos/types.go
@@ -31,9 +31,9 @@ import (
 
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
 	"github.com/chaos-mesh/chaos-mesh/controllers/config"
+	"github.com/chaos-mesh/chaos-mesh/controllers/utils"
 	"github.com/chaos-mesh/chaos-mesh/controllers/recover"
 	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/pb"
-	"github.com/chaos-mesh/chaos-mesh/pkg/controllerutils"
 	"github.com/chaos-mesh/chaos-mesh/pkg/events"
 	"github.com/chaos-mesh/chaos-mesh/pkg/finalizer"
 	"github.com/chaos-mesh/chaos-mesh/pkg/router"
@@ -135,7 +135,7 @@ func (r *endpoint) Recover(ctx context.Context, req ctrl.Request, chaos v1alpha1
 func (r *recoverer) RecoverPod(ctx context.Context, pod *v1.Pod, somechaos v1alpha1.InnerObject) error {
 	r.Log.Info("Try to recover pod", "namespace", pod.Namespace, "name", pod.Name)
 
-	daemonClient, err := controllerutils.NewChaosDaemonClient(ctx, r.Client, pod)
+	daemonClient, err := utils.NewChaosDaemonClient(ctx, r.Client, pod)
 	if err != nil {
 		r.Log.Error(err, "get chaos daemon client")
 		return err
@@ -192,7 +192,7 @@ func (r *endpoint) applyAllPods(ctx context.Context, pods []v1.Pod, chaos *v1alp
 func (r *endpoint) applyPod(ctx context.Context, pod *v1.Pod, dnsServerIP string) error {
 	r.Log.Info("Try to apply dns chaos", "namespace",
 		pod.Namespace, "name", pod.Name)
-	daemonClient, err := controllerutils.NewChaosDaemonClient(ctx, r.Client, pod)
+	daemonClient, err := utils.NewChaosDaemonClient(ctx, r.Client, pod)
 	if err != nil {
 		r.Log.Error(err, "get chaos daemon client")
 		return err
